Set README title.
This is the first manual commit to the new repository I've created for 2025 because I like things clean and I can do whatever I want. Subsequent commits will repopulate this repository with things that were in the previous one, cleaned up and updated to varying degrees as I deem necessary, as well as new things.